move store interface to separate pkg
